merge rev 895 of cin2cv svn: ------------------------------------------------------------------------ r895 | j6t | 2006-09-22 22:21:46 +0200 (Fri, 22 Sep 2006) | 15 lines
Fix clipboard handling to return only data of the requested type.

The modifications look more voluminous than they are because one change
is to move the handler for SelectionRequest events into a set of
functions.

The major fix is to check the type of data that is requested in the
SelectionRequest event. This way a bad interaction with Klipper is
avoided, which requests data of type 'TARGETS' when it notices that the
selection owner has changed.

In addition, the request for 'TARGETS' is also honored. This makes
Cinelerra a little more compliant with the ICCCM. However, the mandatory
targets 'TIMESTAMP' and 'MULTIPLE' are still not supported.

------------------------------------------------------------------------