import rev 693 of cin cv: ------------------------------------------------------------------------ r693 | pmdumuid | 2005-11-22 09:34:45 +0100 (Tue, 22 Nov 2005) | 1 line
Widget Grid system - This system allows you to place widgets in a grid (ignoring the width of the individual widgets) then a call to move_widgets will work out all the widths and spacing for that grid.  This fixes all the language overlapping problem.  Note: This is only the 'engine', all the files that create dialogs need to be modified to make use this tool.
------------------------------------------------------------------------